refactor: separate module bitmap tag translation
